+++ Runtime/HiraScriptCreator/Blackboard/Queries/IIndividualQuery.cs
@@ -3,6 +3,6 @@
     public interface IIndividualQuery
     {
         string Condition { get; }
-        float Weight { get; }
+        int Weight { get; }
     }
 }
